Removes unused metrics
The HttpStart and HttpStop envelope types have been unused and deprecated in loggregator
for a very long time.  We haven't been emitting them for quite a while, and we're not
even using them internally since November, 2015.  We're starting to remove all related
code, so it makes sense to drop them from the docs, as well.

[#126660411]

Signed-off-by: Sam Nelson <snelson@pivotal.io>