Removes information about cpu entitlement plugin
Removes outdated instructions for the CPU Entitlement plugin.

CAPI now exposes a `cpu_entitlement` metric in place of the
`absolute_entitlement` and `absolute_usage` metrics. These will be
captured in the cf CLI soon, but in the meantime the CPU Entitlement
plugin should no longer be the go-to method for retrieving CPU
Entitlement for apps.

Signed-off-by: Rebecca Roberts <rebecca.roberts@broadcom.com>
Signed-off-by: Carson Long <12767276+ctlong@users.noreply.github.com>